backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
psycopg2-binary==2.9.7
SQLAlchemy>=1.4,<3.0
pydantic==1.10.12
passlib[bcrypt]==1.7.4

openpyxl
gunicorn==20.1.0
email-validator==1.3.1
PyJWT>=2.0.0
